+++ db/src/commonMain/kotlin/pw/binom/db/Connection.kt
@@ -5,5 +5,7 @@ import pw.binom.io.Closeable
 interface Connection : Closeable {
     fun createStatement(): Statement
     fun prepareStatement(query: String): PreparedStatement
+    fun commit()
+    fun rollback()
     val type: String
 }