+++ history.md
@@ -1,5 +1,9 @@
 # History
 
+### 0.1.22
+* Fix bugs in utf8 encoder
+* SQLite Autocommit fix bugs
+
 ### 0.1.20
 * Added charset support
 * Added BytesBitset